test: Split out real unit tests
Change-type: patch
