+++ test/integration/worker/utils.spec.js
@@ -21,70 +21,6 @@ const utils = require('../../../lib/worker/utils')
 ava.beforeEach(helpers.jellyfish.beforeEach)
 ava.afterEach(helpers.jellyfish.afterEach)
 
-ava('.getActionArgumentsSchema() should return a wildcard schema if no args', (test) => {
-	const schema = utils.getActionArgumentsSchema({
-		data: {
-			arguments: {}
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object'
-	})
-})
-
-ava('.getActionArgumentsSchema() should parse one argument', (test) => {
-	const schema = utils.getActionArgumentsSchema({
-		data: {
-			arguments: {
-				foo: {
-					type: 'object'
-				}
-			}
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'object'
-			}
-		},
-		additionalProperties: false,
-		required: [ 'foo' ]
-	})
-})
-
-ava('.getActionArgumentsSchema() should parse two arguments', (test) => {
-	const schema = utils.getActionArgumentsSchema({
-		data: {
-			arguments: {
-				foo: {
-					type: 'object'
-				},
-				bar: {
-					type: 'number'
-				}
-			}
-		}
-	})
-
-	test.deepEqual(schema, {
-		type: 'object',
-		properties: {
-			foo: {
-				type: 'object'
-			},
-			bar: {
-				type: 'number'
-			}
-		},
-		additionalProperties: false,
-		required: [ 'foo', 'bar' ]
-	})
-})
-
 ava('.hasCard() id = yes (exists), slug = yes (exists)', async (test) => {
 	const card = await test.context.jellyfish.insertCard(test.context.context, test.context.session, {
 		slug: 'foo-bar',
