+++ test/unit/ui/services/helpers.spec.js
@@ -19,7 +19,7 @@ require('ts-node').register()
 const ava = require('ava')
 const helpers = require('../../../../lib/ui/services/helpers')
 
-ava.only('.createPrefixRegExp() match underscore characters', (test) => {
+ava('.createPrefixRegExp() match underscore characters', (test) => {
 	const matchRE = helpers.createPrefixRegExp('@')
 	const match = matchRE.exec('Lorem ipsum @user_name dolor sit amet')
